Create Perfect numbers in range
Given two integers a and b, write a program to find and print all perfect numbers in the range from a to b.
A perfect number is a positive integer that is equal to the sum of its proper divisors (excluding itself).

Note:-
This is a functional problem. You do not need to take any input. You just need to complete the function, and print the output.
Input
First Line will contain an integer a representing the start of the range.
Second Line will contain an integer b representing the end of the range.
Output
Print all perfect numbers between a and b in a space-separated matter