Create list.to.do.py
This Python code defines a simple to-do list function that stores and displays a set of common tasks in a list format. It initializes a list called `tasks` containing everyday activities such as grocery shopping, finishing work projects, exercising, calling a friend, reading a book, and cleaning the house. When the `todo_list` function is called, it prints each task in the list, providing an easy way for users to keep track of their daily responsibilities.